Add get() for /api/posts/ by ID
I had an issue where it was giving me a status code 200 even if the ID
didn't exists. Had to add .first() to the findById() function in posts_model.js
in order to get it to correctly return the status code 404 with the correct
error message if the ID wasn't found.