Add get() for /api/posts for comments by post ID
Had to add .first() to the end of the findPostComments() function to get it to give the
correct error message. For some reason it is giving the error message with Post ID 9 even
though there is a post with that ID. Need to look into whether there are comments on that post.